Sorter worked on
Sorter isn't working; the page won't load, probably due to an error in a
while loop. "Sorter.html" is the actual project. "merge sort javascript"
is the file which contains the working automatic javascript, but when I
tried to make it a manual sort in "Sorter.html", it didn't work